Refactor logging configuration into its own module
Otherwise we'll have a circular import with main. Trying to pull things
from main.py was sloppy anyways.
